Update after 12/12 work
Clarifying variables, found luminosity. Lum to large so need to look at energy of the particles as well as the radii of the particles. We might have missed a transition from cm to m. Also, need to check units on lum bc that might be part of the ^39 vs ^79 prob that we have.